requirements.txt
# Fabgaurd-AI Requirements
# Python 3.8+ required

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science & ML
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Model Export & Optimization
onnx>=1.12.0
onnxruntime>=1.12.0
# onnxruntime-gpu>=1.12.0  # Uncomment for GPU support in ONNX

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress Bars
tqdm>=4.62.0

# Web API (Optional)
# flask>=2.0.0  # Uncomment if building REST API
# flask-cors>=3.0.0

# Testing (Optional)
# pytest>=7.0.0
# pytest-cov>=3.0.0

# Code Quality (Optional)
# black>=22.0.0
# flake8>=4.0.0
# pylint>=2.12.0
